backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.0
psycopg2-binary==2.9.11

django-crispy-forms==2.3
crispy-bootstrap5==2024.2
geopy==2.4.1
requests==2.32.3
reportlab==4.2.2
python-dotenv>=1.0
channels>=4.0
channels-redis>=4.0.0
celery>=5.4.0
redis==6.4.0
numpy==2.3.3
statsmodels>=0.14.5
scikit-learn>=1.7.2
django-model-utils>=3.1.0
jsonfield>=2.1.0
pytz
swapper
daphne==4.0.0
qrcode[pil]==7.4.2

plotly==5.17.0
pandas>=2.1.4
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
pyzbar>=0.1.9
opencv-python>=4.8.0
barcode>=0.1.0
django-celery-results>=2.5.1



